requirements.txt
bcrypt==3.1.4
certifi==2018.1.18
click==6.7
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
lazy==1.3
MarkupSafe==1.1.1
Pillow==8.0.0
pycparser==2.18
requests==2.18.4
SQLAlchemy==1.2.6
SQLAlchemy-Utils==0.33.2
urllib3==1.22
URLObject==2.4.3
Werkzeug==0.14.1
WTForms==2.1
